+++ src/world/GroupHandler.cpp
@@ -152,7 +152,7 @@ void WorldSession::HandleGroupAcceptOpcode(WorldPacket& recv_data)
 
     // If we're this far, it means we have no existing group, and have to make one.
     grp = new Group(true);
-    grp->m_difficulty = static_cast<uint8>(player->iInstanceType);
+    grp->m_difficulty = player->iInstanceType;
     grp->AddMember(player->m_playerInfo);        // add the inviter first, therefore he is the leader
     grp->AddMember(_player->m_playerInfo);    // add us.
     _player->iInstanceType = grp->m_difficulty;
