+++ src/world/WorldCreator.h
@@ -61,7 +61,7 @@ class SERVER_DECL Instance
         uint32 m_creatorGuid;
         uint32 m_creatorGroup;
         bool m_persistent;
-        uint32 m_difficulty;
+        uint8 m_difficulty;
         std::set<uint32> m_killedNpcs;
         time_t m_creation;
         time_t m_expiration;
